Added the final touch

+++ src/graph.hpp
@@ -71,7 +71,7 @@ class Graph {
     std::mt19937 gen(rd());
  
     std::exponential_distribution<> d(1);
-		for(int i = 0; i < n; i++) {
+		for(uint64_t i = 0; i < n; i++) {
 			for (size_t j = offset[i]; j < offset[i + 1]; j++) {
 				edge[j].w = (int)(10000.0*d(gen));
 			}
